Copy Movies Fan App part 3 into Movies Fan App Part 4
In this lab, we will complete the Movies Fan app.

The only remaining feature is to enable the user to write their review of a favourite movie. Also, there are known bugs identified in the previous lab that we will fix, and we will improve the app's performance by caching the server state.

The new React features used in this development iteration include:

a- The Render props pattern.
b- The Context API.
c- 3rd party hooks, e.g. useForm.
[Note. in the previous lab, you developed an Upcoming movies page as an exercise. We will temporally break this feature as a result of refactoring done in this lab. However, the exercise at the end will allow you to get it working again.] ]