Server state Caching -
Should each component that requires data from the backend/API be responsible for its data fetching?

    a- Ideally, yes, but it should not call the API every time the component is remounted, particularly if the API data changes infrequently.
    b- No, if the component must share the data with other sibling components.

Should we centralizing an app's data management in a high-level component and use props to pass it to subordinates?

      Yes, but only when many components use the data - shared data.

The problem with re-requesting API data every time a component remounts can be solved by caching. When a component (re)mounts with caching enabled, it first checks the local cache for its API data. If the data is there, it uses it; otherwise, it makes an API call. The react-query library is an excellent implementation of this caching behaviour.