Global Data and Contexts -  "Refactor: Favourite movie list now stored in a context."
A common feature of React apps is where many components need to access the same piece of data. For example, many components may need to know the identity of the currently authenticated user. However, this situation can lead to 'prop drilling', where components in the middle of the component hierarchy must pass props from a high-level component to low-level components - the middle-tier components may not use the props themselves. React provides Contexts to solve this problem - it is an implementation of the generic Provider design pattern. We will use a Context to store information about movies tagged as favourites. The Home page and Favourites page will use the Context as both use this data.

Note, TMDB returns a different object structure for a movie when asked for a list of movies rather than details of a particular movie. For this reason, the genre filtering condition is different for the Favourites page as opposed to the home page.

In the browser, try the following sequence:

 1- Go to the home page.

 2- Click the 'Favorites' icon of the first three movies - they each have a favourite icon coloured red in the title.

 3- Navigate to the Favorites page. It's showing the tagged movies.

 4- Navigate back to the Home page. The three movies are still tagged as favourites.