Bug Fix
Currently, the favourite movies page has a bug: the movies listed on the page have a 'Add to Favorites' icon, which is illogical. We will replace it with two options:

 1- Remove from favourites.
 2- Write a review of the movie.

The above code means that movie cards on the Favorite movies page should have two action icons: 'Remove from the Favourites' and 'Write a Review'. In the browser, try adding and removing favourites.