Add a Movie Review
On the favourite movies page, clicking a movie's RateReview icon should display a page that allows the user to write a review. We are storing reviews in the moviesContext - in a production app, it would be sent to the API for storage in a database.

Now when the user clicks the submit button, the snackbar acknowledgement displays. Clicking its close icon will redirect the browser back to the favourites page.